RegIntelAI/requirements.txt
streamlit>=1.29.0
openai>=1.6.1
chromadb>=0.4.18
langchain>=0.1.0
langchain-openai>=0.0.5
pypdf>=3.17.0
python-dotenv>=1.0.0
tiktoken>=0.5.0
